create small objects ingest and broadcast
