+++ release/benchmarks/object_store/small_objects.yaml
